Printing problem resolved
