+++ src/components/Home.jsx
@@ -9,7 +9,7 @@ function Home(){
         user_id:"4EulioOrxqQLhj2n1XDOFlBv7fQ2",
         client_id:"example1",
         timestamp:"2020-11-10 23:59:59",
-        org_id:"South 24 Parganas.8B.JU Hostel.Ekdalia",
+        org_id:"South 24 Parganas.8B.JU Hostel.Office",
         days:10
     });
